Renamed Response to ResponseBase to avoid overlap
