--- backend/src/responses/ErrorResponse.ts
@@ -1,4 +1,0 @@
-export default interface ErrorResponse{
-	errcode: string;
-	message: string;
-}
