+++ backend/src/responses/ResponseBase.ts
@@ -1,0 +1,7 @@
+export enum GenErr{
+	SQL = -1,
+}
+
+export default interface ResponseBase{
+	reason?: string;
+}
